# sampleSource/src/main/resources/META-INF/openapi.json
{
    "swagger": "2.0",
    "info": {
        "title": "test",
        "version": "1.0.0",
        "description": "test",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "petstore.swagger.io",
    "basePath": "/v2",
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/store/inventory": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "format": "int32",
                                "type": "integer"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": [
                        ]
                    }
                ],
                "operationId": "getInventory",
                "summary": "Returns pet inventories by status",
                "description": "Returns a map of status codes to quantities"
            }
        },
        "/store/order": {
            "post": {
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "description": "order placed for purchasing the pet",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        },
                        "in": "body",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    },
                    "400": {
                        "description": "Invalid Order"
                    }
                },
                "operationId": "placeOrder",
                "summary": "Place an order for a pet",
                "description": ""
            }
        },
        "/store/order/{orderId}": {
            "get": {
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "parameters": [
                    {
                        "name": "orderId",
                        "description": "ID of pet that needs to be fetched",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int64",
                        "maximum": 10,
                        "minimum": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "operationId": "getOrderById",
                "summary": "Find purchase order by ID",
                "description": "For valid response try integer IDs with value >= 1 and <= 10. Other values will generated exceptions"
            },
            "delete": {
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "parameters": [
                    {
                        "name": "orderId",
                        "description": "ID of the order that needs to be deleted",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "int64",
                        "minimum": 1
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "operationId": "deleteOrder",
                "summary": "Delete purchase order by ID",
                "description": "For valid response try integer IDs with positive integer value. Negative or non-integer values will generate API errors"
            }
        },
        "/user": {
            "post": {
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "description": "Created user object",
                        "schema": {
                            "$ref": "#/definitions/User"
                        },
                        "in": "body",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "operationId": "createUser",
                "summary": "Create user",
                "description": "This can only be done by the logged in user."
            }
        },
        "/user/createWithArray": {
            "post": {
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "description": "List of user object",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        },
                        "in": "body",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "operationId": "createUsersWithArrayInput",
                "summary": "Creates list of users with given input array",
                "description": ""
            }
        },
        "/user/createWithList": {
            "post": {
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "description": "List of user object",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        },
                        "in": "body",
                        "required": true
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "operationId": "createUsersWithListInput",
                "summary": "Creates list of users with given input array",
                "description": ""
            }
        },
        "/user/login": {
            "get": {
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "description": "The user name for login",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "description": "The password for login in clear text",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "type": "integer",
                                "format": "int32"
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                },
                "operationId": "loginUser",
                "summary": "Logs user into the system",
                "description": ""
            }
        },
...
        "/user/{username}": {
            "get": {
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "description": "The name that needs to be fetched. Use user1 for testing. ",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "operationId": "getUserByName",
                "summary": "Get user by user name",
                "description": ""
            },
            "put": {
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "description": "name that need to be updated",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "description": "Updated user object",
                        "schema": {
                            "$ref": "#/definitions/User"
                        },
                        "in": "body",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid user supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "operationId": "updateUser",
                "summary": "Updated user",
                "description": "This can only be done by the logged in user."
            },
            "delete": {
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "description": "The name that needs to be deleted",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid username supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "operationId": "deleteUser",
                "summary": "Delete user",
                "description": "This can only be done by the logged in user."
            }
        },
...
    },
    "definitions": {
        "Order": {
            "type": "object",
            "properties": {
                "id": {
                    "format": "int64",
                    "type": "integer"
                },
                "petId": {
                    "format": "int64",
                    "type": "integer"
                },
                "quantity": {
                    "format": "int32",
                    "type": "integer"
                },
                "shipDate": {
                    "format": "date-time",
                    "type": "string"
                },
                "status": {
                    "description": "Order Status",
                    "enum": [
                        "placed",
                        "approved",
                        "delivered"
                    ],
                    "type": "string"
                },
                "complete": {
                    "default": false,
                    "type": "boolean"
                }
            },
            "xml": {
                "name": "Order"
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "format": "int64",
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "userStatus": {
                    "format": "int32",
                    "description": "User Status",
                    "type": "integer"
                }
            },
            "xml": {
                "name": "User"
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "flow": "implicit",
            "authorizationUrl": "https://petstore.swagger.io/oauth/dialog",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            },
            "type": "oauth2"
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
...
}